finished functionality to disable on certain pages (#120, #133); * localization complete * added confirmation dialog to options reset * switched filter table delete all to use confirm package for consistency
