+++ package.json
@@ -31,6 +31,7 @@
     "lodash-es": "^4.17.15",
     "match-sorter": "^4.1.0",
     "material-table": "^1.57.2",
+    "material-ui-confirm": "^2.0.4",
     "notistack": "^0.9.11",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
